# server/swagger/mastermind-v3.json
{
    "swagger": "2.0",
    "info": {
        "title": "MasterMind Middleware API",
        "description": "The data storage middleware for MasterMind",
        "version": "3.0.0"
    },
    "host": "localhost:3000",
    "schemes": [
        "http"
    ],
    "basePath": "/v3",
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/departments": {
            "x-handler": "Departments",
            "get": {
                "summary": "Retrieve Departments",
                "description": "The Departments endpoint returns departments.\n",
                "operationId": "getDepartments",
                "tags": [
                    "Departments"
                ],
                "parameters": [
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Filter departments to only those with this code",
                        "type": "string"
                    },
                    {
                        "name": "manager",
                        "in": "query",
                        "description": "Filter departments to only those with the given manager, identified by ID",
                        "type": "string"
                    },
                    {
                        "name": "nickname",
                        "in": "query",
                        "description": "Filter departments to only those whose nickname contains the given string",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An array of departments",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Department"
                            }
                        }
                    },
                    "403": {
                        "description": "The current user is not allowed to view Departments",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "summary": "Create Department",
                "description": "Create a new department.\nAn ID in the POST-ed department will be ignored.\n",
                "operationId": "createSingleDepartment",
                "tags": [
                    "Departments"
                ],
                "parameters": [
                    {
                        "name": "department",
                        "in": "body",
                        "description": "Data representing the department to create",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Department"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The newly created department",
                        "schema": {
                            "$ref": "#/definitions/Department"
                        }
                    },
                    "403": {
                        "description": "The current user is not allowed to create new Departments",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/departments/{id}": {
            "x-handler": "Departments",
            "get": {
                "summary": "Retrieve Department",
                "description": "The Department endpoint returns a single department.\n",
                "operationId": "getSingleDepartment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the department to retrieve.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Departments"
                ],
                "responses": {
                    "200": {
                        "description": "The single department specified by the ID",
                        "schema": {
                            "$ref": "#/definitions/Department"
                        }
                    },
                    "403": {
                        "description": "The current user is not allowed to view Departments",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "A department with the specified ID could not be found.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "summary": "Update Department",
                "description": "Updates an existing department.\n",
                "operationId": "updateSingleDepartment",
                "tags": [
                    "Departments"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the department to update.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "department",
                        "in": "body",
                        "description": "New data representing the department",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Department"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The newly updated department",
                        "schema": {
                            "$ref": "#/definitions/Department"
                        }
                    },
                    "403": {
                        "description": "The current user is not allowed to update Departments",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "A department with the specified ID could not be found.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete Department",
                "description": "Deletes a single Department.\n",
                "operationId": "deleteSingleDepartment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the department to delete.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Departments"
                ],
                "responses": {
                    "200": {
                        "description": "The single department specified by the ID was deleted"
                    },
                    "403": {
                        "description": "The current user is not allowed to delete Departments",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "A department with the specified ID could not be found.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Department": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "category": {
                    "type": "string",
                    "description": "ID of Department Category"
                },
                "code": {
                    "type": "string",
                    "description": "The department code"
                },
                "manager": {
                    "type": "string",
                    "description": "ID of Person"
                }
            },
            "required": [
                "code",
                "nickname",
                "category"
            ]
        },
...
        "Error": {
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32"
                },
                "error": {
                    "type": "string"
                },
                "fields": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "oauth2": {
            "type": "oauth2",
            "authorizationUrl": "http://localhost:3000/auth/google",
            "flow": "implicit",
            "scopes": {
                "https://www.googleapis.com/auth/plus.login": "authenticated via Google Plus"
            }
        }
    },
    "security": [
        {
            "oauth2": [
                "https://www.googleapis.com/auth/plus.login"
            ]
        }
    ]
}